# OpenAPI.yaml
openapi: 3.0.3
info:
  title: Mini-Blog API
  version: 1.0.0
  description: API REST para autores y posts.
  contact:
    name: Sam Bracamonte
    email: sam.bracamonte3@gmail.com

servers:
  - url: https://proyectom2sambracamonte-production.up.railway.app
    description: Desarrollo deployed railway

tags:
  - name: Authors
    description: Endpoints de autores
  - name: Posts
    description: Endpoints de posts

paths:
  /api/authors:
    get:
      summary: Obtener todos los autores
      tags: [Authors]
      parameters: []
      responses:
        "200":
          description: Lista de autores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Author"

    post:
      summary: Crear autor
      tags: [Authors]
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorCreate"
      responses:
        "201":
          description: Autor creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"


  /api/authors/{id}:
    get:
      summary: Obtener autor por id
      tags: [Authors]
      parameters:
        - $ref: "#/components/parameters/AuthorId"
      responses:
        "200":
          description: Autor encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Actualizar autor
      tags: [Authors]
      parameters:
        - $ref: "#/components/parameters/AuthorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorUpdate"
      responses:
        "202":
          description: Autor actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

    delete:
      summary: Eliminar autor
      tags: [Authors]
      parameters:
        - $ref: "#/components/parameters/AuthorId"
      responses:
        "200":
          description: Autor eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/posts:
    get:
      summary: Listar posts
      tags: [Posts]
      parameters:
        - in: query
          name: published
          schema:
            type: boolean
          required: false
          description: Filtra por estado de publicacion
      responses:
        "200":
          description: Lista de posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

    post:
      summary: Crear post
      tags: [Posts]
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostCreate"
      responses:
        "201":
          description: Post creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/posts/author/{authorId}:
    get:
      summary: Listar posts por autor
      tags: [Posts]
      parameters:
        - $ref: "#/components/parameters/AuthorIdPath"
      responses:
        "200":
          description: Lista de posts del autor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/posts/{id}:
    get:
      summary: Obtener post por id
      tags: [Posts]
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Post encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Actualizar post
      tags: [Posts]
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Post actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Eliminar post
      tags: [Posts]
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Post eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  schemas:

    Author:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        bio:
          type: string
          nullable: true
      required: [id, name, email]

    AuthorCreate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        bio:
          type: string
      required: [name, email]

    AuthorUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        bio:
          type: string

    Post:
      type: object
      properties:
        id:
          type: integer
        author_id:
          type: integer
        title:
          type: string
        content:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, author_id, title, content, published, created_at]

    PostCreate:
      type: object
      properties:
        author_id:
          type: integer
        title:
          type: string
        content:
          type: string
        published:
          type: boolean
      required: [author_id, title, content]
    PostUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        published:
          type: boolean

    Message:
      type: object
      properties:
        message:
          type: string
      required: [message]
      
    Error:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer
      required: [error, status]
  parameters:
    AuthorId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    AuthorIdPath:
      name: authorId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    PostId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Solicitud invalida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflicto de unicidad
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"